Removed unused components and unnessesory console.log

+++ src/app/Category/category-list/category-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { CategoryService } from 'src/app/Services/CategoryService/category.service';
-import { ActivatedRoute, Router } from '@angular/router';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-category-list',
@@ -11,40 +11,42 @@ export class CategoryListComponent implements OnInit {
   public allCategoryData:any;
   public status:number = null;
 
-  constructor(public categoryService:CategoryService, public router:Router) { 
-    console.log("Category Constructor");
-  }
+  constructor(public categoryService:CategoryService, public router:Router) { }
 
   ngOnInit() {
+    this.status=0;
     this.getCategoryData();
   }
 
   public getCategoryData:any = ()=>{
+    this.allCategoryData = null;
     this.categoryService.getAllData().subscribe(
       data=>{
-        console.log(data);
+        this.status=1;
         this.allCategoryData = data;
       },
       error=>{
-        console.log("Error in getAllData:- ");
+        this.status = -1;
         console.log(error.errorMessage);
       }
     )
   }
 
   public deleteCategory:any = (id:number)=>{
-    this.categoryService.deleteData(id).subscribe(
-      data=>{
-        console.log("Success Deletion");
-        setTimeout(()=>{
-          this.router.navigate(['/categorylist']);
-        },2000);
-      },
-      error=>{
-        console.log("Error Deletion:- ");
-        console.log(error.errorMessage);
-      }
-    )
+    let choice = confirm("Are you sure want to Delete?");
+    if(choice){
+      this.categoryService.deleteData(id).subscribe(
+        data=>{
+          alert("Delete Successful!");
+            this.getCategoryData();
+        },
+        error=>{
+          console.log(error.errorMessage);
+          alert("Sorry you can't Delete This Category!");
+        }
+      )
+    }
+    
   }
 
 }
